chore(typescript): migrated `victory-line` to TS
