--- packages/victory-line/src/index.js
@@ -1,2 +1,0 @@
-export { default as VictoryLine } from "./victory-line";
-export { default as Curve } from "./curve";
