+++ packages/victory-line/src/index.ts
@@ -1,0 +1,2 @@
+export * from "./victory-line";
+export * from "./curve";
